+++ src/layout/column/index.tsx
@@ -5,6 +5,9 @@ import Column from './column'
 const StyledColumn = styled(Column)`
   & {
     flex: 1;
+    text-indent: 0;
+    margin: 0;
+    padding: 0;
   }
 `