+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Unreleased
 
+- [...]
+
+# v21.2.0 (20/02/2020)
+
 - **[NEW]** Introduce `LayoutNormalizer` with legacy global layout overrides. Not activated by default.
 - **[NEW]** Add Bus Ride Details story page
 - **[NEW]** Implement `ColumnedContentSection` widget.
 - **[UPDATE]** Use semantic html (ul/li) for `Columns` and `Column` widgets.
 - **[UPDATE]** Add new `Your rides`, `Ride history` and `Search result` pages
 - **[UPDATE]** Make price props on `TripCard` non-mandatory
-- [...]
 
 # v21.1.0 (19/02/2020)